Remove unused components and update imports
