+++ src/features/cashier-dashboard/components/cashier_dashboard_order_details_sidebar/OrderDetailsSideBar.cy.tsx
