+++ src/features/cashier-dashboard/components/cashier_dashboard_payement_sidebar/PaymentDrawer.tsx
@@ -1,28 +1,24 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import { RxCross1 } from 'react-icons/rx';
-import HorizontalDivider from '../../shared/divider/HorizontalDivider';
+import HorizontalDivider from '../../../../shared/divider/HorizontalDivider';
 import { BiReceipt } from 'react-icons/bi';
 import { AiOutlineMail } from 'react-icons/ai';
 import { MdOutlineDone } from 'react-icons/md';
-import PaymentMethodButton from '../../shared/buttons/PaymentMethodButton';
-import PaymentNumberPad from '../../shared/payment_numberpad/PaymentNumberPad';
-import FooterButton from '../../shared/buttons/FooterButton';
-import mastercard from '../../assets/icons/mastercard.png';
-import money from '../../assets/icons/money.png';
-import visa from '../../assets/icons/visa.png';
+import PaymentMethodButton from '../../../../shared/buttons/PaymentMethodButton';
+import PaymentNumberPad from '../../../../shared/payment_numberpad/PaymentNumberPad';
+import FooterButton from '../../../../shared/buttons/FooterButton';
+import mastercard from '../../../../assets/icons/mastercard.png';
+import money from '../../../../assets/icons/money.png';
+import visa from '../../../../assets/icons/visa.png';
 import {
   ComponentState,
   usePaymentContext,
-} from '../../features/cashier-dashboard/layout/MainCashierDashboard';
-type Props = {
-  openConfirmPayment: () => void;
-};
+} from '../../layout/MainCashierDashboard';
 
 const PaymentDrawer = () => {
   const { setCurrentComponent, paymentDetails, setPaymentDetails } =
     usePaymentContext();
 
-  const handlePaymentMethodClick = () => {};
   const [payment, setPayment] = useState('');
   const handleKeyPress = (key: string) => {
     if (key === 'C') {
