+++ src/features/cashier-dashboard/interfaces/OrderMedicine.tsx
@@ -1,0 +1,7 @@
+export interface OrderedMedicine {
+  id: string;
+  name: string;
+  unitPrice: number;
+  amount: number;
+  availableQuantity: number;
+}
