+++ src/features/cashier-dashboard/interfaces/PaymentDetails.tsx
@@ -1,0 +1,9 @@
+export interface PaymentDetails {
+  paymentMethod: string;
+  paymentAmount: number;
+  paymentDate: Date;
+  paymentType: string;
+  paymentNotes: string;
+  paymentDiscount: number;
+  paidAmount: number;
+}
